move time-related library code to dedicated files
