+++ include/h2o.h
@@ -41,6 +41,7 @@ extern "C" {
 #include "h2o/memory.h"
 #include "h2o/socket.h"
 #include "h2o/string_.h"
+#include "h2o/time_.h"
 #include "h2o/timeout.h"
 #include "h2o/version.h"
