+++ include/h2o/string_.h
@@ -36,9 +36,6 @@ extern "C" {
 
 #define H2O_STRLIT(s) (s), sizeof(s) - 1
 
-#define H2O_TIMESTR_RFC1123_LEN (sizeof("Sun, 06 Nov 1994 08:49:37 GMT") - 1)
-#define H2O_TIMESTR_LOG_LEN (sizeof("29/Aug/2014:15:34:38 +0900") - 1)
-
 /**
  * duplicates given string
  * @param pool memory pool (or NULL to use malloc)
@@ -75,14 +72,6 @@ h2o_iovec_t h2o_decode_base64url(h2o_mem_pool_t *pool, const char *src, size_t l
  * base64 encoder
  */
 void h2o_base64_encode(char *dst, const void *src, size_t len, int url_encoded);
-/**
- * builds a RFC-1123 style date string
- */
-void h2o_time2str_rfc1123(char *buf, struct tm *gmt);
-/**
- * builds an Apache log-style date string
- */
-void h2o_time2str_log(char *buf, time_t time);
 /**
  * returns the extension portion of path
  */
